+++ src/main/java/kitchenpos/domain/order/OrderRepository.java
@@ -1,8 +1,7 @@
-package kitchenpos.dao;
+package kitchenpos.domain.order;
 
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
