+++ src/test/java/kitchenpos/application/menugroup/MenuGroupServiceTest.java
@@ -1,10 +1,9 @@
-package kitchenpos.application;
+package kitchenpos.application.menugroup;
 
 import static kitchenpos.fixture.MenuGroupFixture.메뉴_그룹;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.application.menugroup.MenuGroupService;
 import kitchenpos.domain.menugroup.MenuGroup;
 import kitchenpos.domain.menugroup.MenuGroupRepository;
 import kitchenpos.dto.menugroup.MenuGroupRequest;
